Add space archetype in platformer scene
Implemented a new archetype named 'space' in the platformer scene of the game. In addition to creating the archetype, a 'resolv.Space' component was also introduced and used to define 'space' within the newly included factory module. This implementation aims to manage the game's world space as a grid of cells for object-collision checks.